Fix warning: the use of tmpnam is dangerous
While compiling the plugin the following warning accure
warning: the use of tmpnam' is dangerous, better use mkstemp'

Because NcFile object only needs a temporary file path for
creating a file and mkstemp also creates a file desctiptor
it is not reasonable to use it in here. So the temporary file
path creation is replaced with UniqueTemporaryPath object.