Use configured metadata url values in ExtendedCapabilities
There is two special keys ("ID" and "constraint") tuery settings for
picking 'uuid' and 'id' values specified in dataset configurations.

If a dataset configuration has dataset id "myid". The setting
{ key = "ID"; } use the uuid value of the dataset. But if the value is
defined in the plugin configuration e.g. { key = "ID"; value = "myid2"; }
"myid2" is used over the dataset ("myid") configuration. The ID can be
used with GetRecordById request of CSW.

If the key is "constraint" and the value is not defined, the value is
picked from the id variable of dataset configuration. The result KVP is
then  "constraint?=ResourceIdentifier=%27myid%27". The constraint can be
used with GetRecords request of CSW.

These rules applies only for the key-value pairs of the query settings.

There is also possible to define constant URL for the metadataURL
attribute of inspire_dls:SpatialDataSetIdentifier.

Structured configuration is not possible to use in
inspire_common:MetadataUrl element with the capabilities.template.